:wrench: De-globalized `BeamFactory` and `CharacterFactory`
RATIONALE:

Objects relevant only to simulation session shouldn't be global (requires complex reset logic and causes bugs), but their lifetime should be tied to a single sim session.
In this case, the sim. session is tied to lifetime of RoRFrameListener object (newly referred to as 'sim controller') and all sim-specific objects are being de-globalized and put under it's management. In the end, RoRFrameListener will be renamed to SimController

REMARKS:

* Removed useless interface IBehaviorManager - only single client, never reasonably used.
* call to `BeamFactory::windowResized()` moved from GUIManager (it's only reference to BeamFactory) to RoRFrameListener. This makes no difference as both are invoked upon the same RoR::WindowEventListener callback.
* Removed tiny function `TerrainManager::initCollisions()` - replaced by plain code.